lib: update ht16k33 keypad code
Poll the internet value before reading the values.  In theory
this is lower overhead.

Was seeing problems where if you held down a key it was repeating rather
than staying down.

Turns out it's just a bad idea to read out the keypad faster than the 20ms
rate that the kt16k33 updates itself.